Only run bundle install when not a dry run
We don't need to `bundle install` if we are simply downloading data.

A dry run only downloads data and is often run outside of the
development VM. Currently, outside the VM, this is somewhat painful
because we are using ruby 1.9.3 and we have to specifically install
an old version of bundler to make this work.

+++ development-vm/replication/replicate-data-local.sh
@@ -8,8 +8,10 @@ set -eu
 . $(dirname $0)/common-args.sh
 . $(dirname $0)/aws.sh
 
-status "Running bundle install"
-bundle install --quiet
+if ! $DRY_RUN; then
+  status "Running bundle install"
+  bundle install --quiet
+fi
 
 $(dirname $0)/sync-aws-mysql.sh "$@" mysql-master
